refactor pkg/appfile (#1447)
