+++ pkg/appfile/parser_test.go
@@ -27,7 +27,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -35,12 +34,9 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/pointer"
 
-	"github.com/oam-dev/kubevela/apis/core.oam.dev/common"
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/v1alpha2"
 	"github.com/oam-dev/kubevela/apis/core.oam.dev/v1beta1"
-	oamtypes "github.com/oam-dev/kubevela/apis/types"
 	"github.com/oam-dev/kubevela/pkg/dsl/definition"
 	"github.com/oam-dev/kubevela/pkg/dsl/process"
 	"github.com/oam-dev/kubevela/pkg/oam"
@@ -48,7 +44,7 @@ import (
 )
 
 var expectedExceptApp = &Appfile{
-	Name: "test",
+	Name: "application-sample",
 	Workloads: []*Workload{
 		{
 			Name: "myweb",
@@ -57,7 +53,7 @@ var expectedExceptApp = &Appfile{
 				"image": "busybox",
 				"cmd":   []interface{}{"sleep", "1000"},
 			},
-			FullTemplate: &util.Template{
+			FullTemplate: &Template{
 				TemplateStr: `
       output: {
         apiVersion: "apps/v1"
@@ -209,6 +205,7 @@ apiVersion: core.oam.dev/v1beta1
 kind: Application
 metadata:
   name: application-sample
+  namespace: default
 spec:
   components:
     - name: myweb
@@ -251,7 +248,7 @@ var _ = Describe("Test application parser", func() {
 			},
 		}
 
-		appfile, err := NewApplicationParser(&tclient, dm, pd).GenerateAppFile(context.TODO(), "test", &o)
+		appfile, err := NewApplicationParser(&tclient, dm, pd).GenerateAppFile(context.TODO(), &o)
 		Expect(err).ShouldNot(HaveOccurred())
 		Expect(equal(expectedExceptApp, appfile)).Should(BeTrue())
 	})
@@ -291,6 +288,7 @@ var _ = Describe("Test appFile parser", func() {
 		var TestApp = &Appfile{
 			RevisionName: "test-v1",
 			Name:         "test",
+			Namespace:    "default",
 			Workloads: []*Workload{
 				{
 					Name: "myweb",
@@ -300,6 +298,10 @@ var _ = Describe("Test appFile parser", func() {
 						"cmd":    []interface{}{"sleep", "1000"},
 						"config": "myconfig",
 					},
+					UserConfigs: []map[string]string{
+						{"name": "c1", "value": "v1"},
+						{"name": "c2", "value": "v2"},
+					},
 					Scopes: []Scope{
 						{Name: "test-scope", GVK: schema.GroupVersionKind{
 							Group:   "core.oam.dev",
@@ -308,7 +310,7 @@ var _ = Describe("Test appFile parser", func() {
 						}},
 					},
 					engine: definition.NewWorkloadAbstractEngine("myweb", pd),
-					FullTemplate: &util.Template{
+					FullTemplate: &Template{
 						TemplateStr: `
       output: {
         apiVersion: "apps/v1"
@@ -382,7 +384,7 @@ var _ = Describe("Test appFile parser", func() {
 			Data:       map[string]string{"c1": "v1", "c2": "v2"},
 		}
 		Expect(k8sClient.Create(context.Background(), cm.DeepCopy())).Should(SatisfyAny(BeNil(), &util.AlreadyExistMatcher{}))
-		ac, components, err := NewApplicationParser(k8sClient, dm, pd).GenerateApplicationConfiguration(TestApp, "default")
+		ac, components, err := TestApp.GenerateApplicationConfiguration()
 		Expect(err).To(BeNil())
 		manuscaler := util.Object2RawExtension(&unstructured.Unstructured{
 			Object: map[string]interface{}{
@@ -499,369 +501,12 @@ var _ = Describe("Test appFile parser", func() {
 		Expect(len(components)).To(BeEquivalentTo(1))
 		Expect(components[0].ObjectMeta).To(BeEquivalentTo(expectComponent.ObjectMeta))
 		Expect(components[0].TypeMeta).To(BeEquivalentTo(expectComponent.TypeMeta))
+		By(string(components[0].Spec.Workload.Raw))
 		Expect(components[0].Spec.Workload).Should(SatisfyAny(
 			BeEquivalentTo(util.Object2RawExtension(expectWorkload)),
 			BeEquivalentTo(util.Object2RawExtension(expectWorkloadOptional))))
 	})
 
-})
-
-var _ = Describe("Test appfile parser to parse helm module", func() {
-	var (
-		appName  = "test-app"
-		compName = "test-comp"
-	)
-
-	It("Test application containing helm module", func() {
-		appFile := &Appfile{
-			Name:         appName,
-			RevisionName: appName + "-v1",
-			Workloads: []*Workload{
-				{
-					Name:               compName,
-					Type:               "webapp-chart",
-					CapabilityCategory: oamtypes.HelmCategory,
-					Params: map[string]interface{}{
-						"image": map[string]interface{}{
-							"tag": "5.1.2",
-						},
-					},
-					engine: definition.NewWorkloadAbstractEngine(compName, pd),
-					Traits: []*Trait{
-						{
-							Name: "scaler",
-							Params: map[string]interface{}{
-								"replicas": float64(10),
-							},
-							engine: definition.NewTraitAbstractEngine("scaler", pd),
-							Template: `
-      outputs: scaler: {
-      	apiVersion: "core.oam.dev/v1alpha2"
-      	kind:       "ManualScalerTrait"
-      	spec: {
-      		replicaCount: parameter.replicas
-      	}
-      }
-      parameter: {
-      	//+short=r
-      	replicas: *1 | int
-      }
-`,
-						},
-					},
-					FullTemplate: &util.Template{
-						Reference: common.WorkloadGVK{
-							APIVersion: "apps/v1",
-							Kind:       "Deployment",
-						},
-						Helm: &common.Helm{
-							Release: util.Object2RawExtension(map[string]interface{}{
-								"chart": map[string]interface{}{
-									"spec": map[string]interface{}{
-										"chart":   "podinfo",
-										"version": "5.1.4",
-									},
-								},
-							}),
-							Repository: util.Object2RawExtension(map[string]interface{}{
-								"url": "http://oam.dev/catalog/",
-							}),
-						},
-					},
-				},
-			},
-		}
-		By("Generate ApplicationConfiguration and Components")
-		ac, components, err := NewApplicationParser(k8sClient, dm, pd).GenerateApplicationConfiguration(appFile, "default")
-		Expect(err).To(BeNil())
-
-		manuscaler := util.Object2RawExtension(&unstructured.Unstructured{
-			Object: map[string]interface{}{
-				"apiVersion": "core.oam.dev/v1alpha2",
-				"kind":       "ManualScalerTrait",
-				"metadata": map[string]interface{}{
-					"labels": map[string]interface{}{
-						"app.oam.dev/component":   compName,
-						"app.oam.dev/name":        appName,
-						"trait.oam.dev/type":      "scaler",
-						"trait.oam.dev/resource":  "scaler",
-						"app.oam.dev/appRevision": appName + "-v1",
-					},
-				},
-				"spec": map[string]interface{}{"replicaCount": int64(10)},
-			},
-		})
-		expectAppConfig := &v1alpha2.ApplicationConfiguration{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "ApplicationConfiguration",
-				APIVersion: "core.oam.dev/v1alpha2",
-			}, ObjectMeta: metav1.ObjectMeta{
-				Name:      appName,
-				Namespace: "default",
-				Labels:    map[string]string{oam.LabelAppName: appName},
-			},
-			Spec: v1alpha2.ApplicationConfigurationSpec{
-				Components: []v1alpha2.ApplicationConfigurationComponent{
-					{
-						ComponentName: compName,
-						Traits: []v1alpha2.ComponentTrait{
-							{
-								Trait: manuscaler,
-							},
-						},
-					},
-				},
-			},
-		}
-		expectComponent := &v1alpha2.Component{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "Component",
-				APIVersion: "core.oam.dev/v1alpha2",
-			}, ObjectMeta: metav1.ObjectMeta{
-				Name:      compName,
-				Namespace: "default",
-				Labels:    map[string]string{oam.LabelAppName: appName},
-			},
-			Spec: v1alpha2.ComponentSpec{
-				Helm: &common.Helm{
-					Release: util.Object2RawExtension(map[string]interface{}{
-						"apiVersion": "helm.toolkit.fluxcd.io/v2beta1",
-						"kind":       "HelmRelease",
-						"metadata": map[string]interface{}{
-							"name":      fmt.Sprintf("%s-%s", appName, compName),
-							"namespace": "default",
-						},
-						"spec": map[string]interface{}{
-							"chart": map[string]interface{}{
-								"spec": map[string]interface{}{
-									"sourceRef": map[string]interface{}{
-										"kind":      "HelmRepository",
-										"name":      fmt.Sprintf("%s-%s", appName, compName),
-										"namespace": "default",
-									},
-								},
-							},
-							"interval": "5m0s",
-							"values": map[string]interface{}{
-								"image": map[string]interface{}{
-									"tag": "5.1.2",
-								},
-							},
-						},
-					}),
-					Repository: util.Object2RawExtension(map[string]interface{}{
-						"apiVersion": "source.toolkit.fluxcd.io/v1beta1",
-						"kind":       "HelmRepository",
-						"metadata": map[string]interface{}{
-							"name":      fmt.Sprintf("%s-%s", appName, compName),
-							"namespace": "default",
-						},
-						"spec": map[string]interface{}{
-							"url": "http://oam.dev/catalog/",
-						},
-					}),
-				},
-				Workload: util.Object2RawExtension(map[string]interface{}{
-					"apiVersion": "apps/v1",
-					"kind":       "Deployment",
-					"metadata": map[string]interface{}{
-						"labels": map[string]interface{}{
-							"workload.oam.dev/type":   "webapp-chart",
-							"app.oam.dev/component":   compName,
-							"app.oam.dev/name":        appName,
-							"app.oam.dev/appRevision": appName + "-v1",
-						},
-					},
-				}),
-			},
-		}
-		By("Verify expected ApplicationConfiguration")
-		diff := cmp.Diff(ac, expectAppConfig)
-		Expect(diff).Should(BeEmpty())
-		By("Verify expected Component")
-		diff = cmp.Diff(components[0], expectComponent)
-		Expect(diff).ShouldNot(BeEmpty())
-	})
-
-})
-
-var _ = Describe("Test appfile parser to parse kube module", func() {
-	var (
-		appName  = "test-app"
-		compName = "test-comp"
-	)
-	var testTemplate = func() runtime.RawExtension {
-		yamlStr := `apiVersion: apps/v1
-kind: Deployment
-spec:
-  selector:
-    matchLabels:
-      app: nginx
-  template:
-    metadata:
-      labels:
-        app: nginx
-    spec:
-      containers:
-      - name: nginx
-      ports:
-      - containerPort: 80 `
-		b, _ := yaml.YAMLToJSON([]byte(yamlStr))
-		return runtime.RawExtension{Raw: b}
-	}
-	var expectWorkload = func() runtime.RawExtension {
-		yamlStr := `apiVersion: apps/v1
-kind: Deployment
-spec:
-  selector:
-    matchLabels:
-      app: nginx
-  template:
-    metadata:
-      labels:
-        app: nginx
-    spec:
-      containers:
-      - name: nginx
-        image: nginx:1.14.0
-      ports:
-      - containerPort: 80 `
-		b, _ := yaml.YAMLToJSON([]byte(yamlStr))
-		return runtime.RawExtension{Raw: b}
-	}
-	var testAppfile = func() *Appfile {
-		return &Appfile{
-			RevisionName: appName + "-v1",
-			Name:         appName,
-			Workloads: []*Workload{
-				{
-					Name:               compName,
-					Type:               "kube-worker",
-					CapabilityCategory: oamtypes.KubeCategory,
-					Params: map[string]interface{}{
-						"image": "nginx:1.14.0",
-					},
-					engine: definition.NewWorkloadAbstractEngine(compName, pd),
-					Traits: []*Trait{
-						{
-							Name: "scaler",
-							Params: map[string]interface{}{
-								"replicas": float64(10),
-							},
-							engine: definition.NewTraitAbstractEngine("scaler", pd),
-							Template: `
-      outputs: scaler: {
-      	apiVersion: "core.oam.dev/v1alpha2"
-      	kind:       "ManualScalerTrait"
-      	spec: {
-      		replicaCount: parameter.replicas
-      	}
-      }
-      parameter: {
-      	//+short=r
-      	replicas: *1 | int
-      }
-`,
-						},
-					},
-					FullTemplate: &util.Template{
-						Kube: &common.Kube{
-							Template: testTemplate(),
-							Parameters: []common.KubeParameter{
-								{
-									Name:       "image",
-									ValueType:  common.StringType,
-									Required:   pointer.BoolPtr(true),
-									FieldPaths: []string{"spec.template.spec.containers[0].image"},
-								},
-							},
-						},
-						Reference: common.WorkloadGVK{
-							APIVersion: "apps/v1",
-							Kind:       "Deployment",
-						},
-					},
-				},
-			},
-		}
-
-	}
-
-	manuscaler := util.Object2RawExtension(&unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "core.oam.dev/v1alpha2",
-			"kind":       "ManualScalerTrait",
-			"metadata": map[string]interface{}{
-				"labels": map[string]interface{}{
-					"app.oam.dev/component":   compName,
-					"app.oam.dev/name":        appName,
-					"app.oam.dev/appRevision": appName + "-v1",
-					"trait.oam.dev/type":      "scaler",
-					"trait.oam.dev/resource":  "scaler",
-				},
-			},
-			"spec": map[string]interface{}{"replicaCount": int64(10)},
-		},
-	})
-
-	It("Test application containing kube module", func() {
-		By("Generate ApplicationConfiguration and Components")
-		ac, components, err := NewApplicationParser(k8sClient, dm, pd).GenerateApplicationConfiguration(testAppfile(), "default")
-		Expect(err).To(BeNil())
-
-		expectAppConfig := &v1alpha2.ApplicationConfiguration{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "ApplicationConfiguration",
-				APIVersion: "core.oam.dev/v1alpha2",
-			}, ObjectMeta: metav1.ObjectMeta{
-				Name:      appName,
-				Namespace: "default",
-				Labels:    map[string]string{oam.LabelAppName: appName},
-			},
-			Spec: v1alpha2.ApplicationConfigurationSpec{
-				Components: []v1alpha2.ApplicationConfigurationComponent{
-					{
-						ComponentName: compName,
-						Traits: []v1alpha2.ComponentTrait{
-							{
-								Trait: manuscaler,
-							},
-						},
-					},
-				},
-			},
-		}
-		expectComponent := &v1alpha2.Component{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "Component",
-				APIVersion: "core.oam.dev/v1alpha2",
-			}, ObjectMeta: metav1.ObjectMeta{
-				Name:      compName,
-				Namespace: "default",
-				Labels:    map[string]string{oam.LabelAppName: appName},
-			},
-			Spec: v1alpha2.ComponentSpec{
-				Workload: expectWorkload(),
-			},
-		}
-		By("Verify expected ApplicationConfiguration")
-		diff := cmp.Diff(ac, expectAppConfig)
-		Expect(diff).Should(BeEmpty())
-		By("Verify expected Component")
-		diff = cmp.Diff(components[0], expectComponent)
-		Expect(diff).ShouldNot(BeEmpty())
-	})
-
-	It("Test missing set required parameter", func() {
-		appfile := testAppfile()
-		// remove parameter settings
-		appfile.Workloads[0].Params = nil
-		_, _, err := NewApplicationParser(k8sClient, dm, pd).GenerateApplicationConfiguration(appfile, "default")
-
-		expectError := errors.WithMessage(errors.New(`require parameter "image"`), "cannot resolve parameter settings")
-		diff := cmp.Diff(expectError, err, test.EquateErrors())
-		Expect(diff).Should(BeEmpty())
-	})
 })
 
 var _ = Describe("Test Get OutputSecretNames", func() {
@@ -929,7 +574,7 @@ settings: {
 
 			wl := &Workload{
 				Name:         "abc",
-				FullTemplate: &util.Template{TemplateStr: template},
+				FullTemplate: &Template{TemplateStr: template},
 			}
 			By("call target function")
 			secrets, err := parseWorkloadInsertSecretTo(ctx, k8sClient, ns, wl)
@@ -969,7 +614,7 @@ parameter: {
 				Params: map[string]interface{}{
 					"dbSecret": targetSecretName,
 				},
-				FullTemplate: &util.Template{TemplateStr: template},
+				FullTemplate: &Template{TemplateStr: template},
 			}
 			By("create secret")
 			s := &corev1.Secret{
@@ -1027,7 +672,7 @@ var _ = Describe("Test IsCloudResourceConsumer", func() {
 	Context("Workload is a Cloud Resource consumer", func() {
 		It("", func() {
 			wl := &Workload{
-				FullTemplate: &util.Template{TemplateStr: "// +insertSecretTo=dbConn"},
+				FullTemplate: &Template{TemplateStr: "// +insertSecretTo=dbConn"},
 			}
 			Expect(wl.IsCloudResourceConsumer()).Should(Equal(true))
 		})
@@ -1036,7 +681,7 @@ var _ = Describe("Test IsCloudResourceConsumer", func() {
 	Context("Workload is a Cloud Resource consumer", func() {
 		It("", func() {
 			wl := &Workload{
-				FullTemplate: &util.Template{TemplateStr: "// +useage=dbConn"},
+				FullTemplate: &Template{TemplateStr: "// +useage=dbConn"},
 			}
 			Expect(wl.IsCloudResourceProducer()).Should(Equal(false))
 		})
