+++ pkg/webhook/core.oam.dev/v1alpha2/application/validation.go
@@ -30,7 +30,7 @@ func (h *ValidatingHandler) ValidateCreate(ctx context.Context, app *v1beta1.App
 	var componentErrs field.ErrorList
 	// try to generate an app file
 	appParser := appfile.NewApplicationParser(h.Client, h.dm, h.pd)
-	if _, err := appParser.GenerateAppFile(ctx, app.Name, app); err != nil {
+	if _, err := appParser.GenerateAppFile(ctx, app); err != nil {
 		componentErrs = append(componentErrs, field.Invalid(field.NewPath("spec"), app, err.Error()))
 	}
 	return componentErrs
